requirements.txt
openai>=1.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0

